Rename org.robockets.robot to org.robockets.tshirtcannon.
Per suggestion of Simon I have renamed org.robockets.robot to
org.robockets.tshirtcannon, this is to avoid future conflicts in the case that
we append the year that the robot is being made to the end of the package name.
