+++ src/org/robockets/tshirtcannon/OI.java
@@ -1,4 +1,4 @@
-package org.robockets.robot;
+package org.robockets.tshirtcannon;
 
 import edu.wpi.first.wpilibj.buttons.Button;
